+++ src/umc-angular-google-analytics.js
@@ -3,65 +3,108 @@
 angular.module('umc-angular-google-analytics', [])
     .provider('Analytics', function() {
         'use strict';
-        var created = false,
-            trackRoutes = true,
-            accountId,
-            trackPrefix = '',
-            domainName,
-            filename = 'analytics.js',
-            pageEvent = '$routeChangeSuccess',
-            trackEcommerce = false,
-            ecommerceLoaded = false;
-
+		
+        var created = false;
+        var trackRoutes = true;
+        var trackPrefix = '';
+        var domainName;
+		var filename = 'analytics.js';
+		var pageEvent = '$routeChangeSuccess';
+		var trackEcommerce = false;
+		var ecommerceLoaded = false;
+		var trackDisplayfeatures = false;
+		var displayfeaturesLoaded = false;
+		var trackEnhancedEcommerce = false;
+		var enhancedEcommerceLoaded = false;
+		
+		this.trackers = [];
         this._logs = [];
 
         // config methods
         this.setAccount = function(id) {
-          accountId = id;
+			// first tracker in array is always the "account" value
+			if (this.trackers.length === 0) {
+				this.trackers.push({ code: id, name: ''});
+			} else {
+				this.trackers[0] = { code: id, name: ''};
+			}
           return true;
         };
+		
         this.trackPages = function(doTrack) {
           trackRoutes = doTrack;
           return true;
         };
+		
         this.trackPrefix = function(prefix) {
           trackPrefix = prefix;
           return true;
         };
 
         this.setDomainName = function(domain) {
-          domainName = domain;
-          return true;
+			domainName = domain;
+			return true;
         };
         
         this.setFilename = function(name) {
-          filename = name;
-          return true;
+			filename = name;
+			return true;
         };
 
         this.setPageEvent = function(name) {
-          pageEvent = name;
-          return true;
+			pageEvent = name;
+			return true;
         };
 
         this.trackEcommerce = function(doTrack) {
-          trackEcommerce = doTrack;
-          return true;
+			trackEcommerce = doTrack;
+			return true;
         };
-
+		
+		this.trackEnhancedEcommerce = function(doTrack) {
+			trackEnhancedEcommerce = doTrack;
+			return true;
+		};
+		
+		this.trackDisplayFeatures = function(doTrack) {
+			trackDisplayfeatures = doTrack;
+			return true;
+		};
+		
+		this.addTracker = function(code, name) {
+			// handle special case of primary tracker
+			if (name === null || name === '') {
+				this.trackers[0].code = code;
+				return;
+			}
+			
+			for (var i = 1; i < this.trackers.length; i++) {
+				if (this.trackers[i].name === name) {
+					this.trackers[i].code = code;
+					return;
+				}
+			} 
+			
+			this.trackers.push({ code: code, name: name });
+		};
+		
+		this.getTrackers = function() {
+			return this.trackers;
+		};
 
         // public service
         this.$get = ['$document', '$rootScope', '$location', '$window', function($document, $rootScope, $location, $window) {
           // private methods
           this._createScriptTag = function() {
-            //require accountId
-            if (!accountId) return;
+            //require a tracking id
+            if (this.trackers.length === 0) 
+            	return;
 
             //initialize the window object __gaTracker
             $window.GoogleAnalyticsObject = '__gaTracker';
             if (angular.isUndefined($window.__gaTracker)) {
                 $window.__gaTracker = function() {
-                    if(angular.isUndefined($window.__gaTracker.q)) {
+                    if (angular.isUndefined($window.__gaTracker.q)) {
                         $window.__gaTracker.q = [];
                     }
                     $window.__gaTracker.q.push(arguments);
@@ -69,15 +112,39 @@ angular.module('umc-angular-google-analytics', [])
                 $window.__gaTracker.l=1*new Date();
             }
             var opts = {};
-            if(domainName) {
+            if (domainName) {
                 opts.cookieDomain = domainName;
             }
 
-            $window.__gaTracker('create', accountId);
+			// create the primary tracker
+            $window.__gaTracker('create', this.trackers[0].code);
+			
+			// create secondary trackers if present
+			for (var i = 1; i < this.trackers.length; i++) {
+				$window.__gaTracker('create', this.trackers[i].code, {'name': this.trackers[i].name });
+			}
 
             if (trackEcommerce && !ecommerceLoaded) {
                 $window.__gaTracker('require', 'ecommerce', 'ecommerce.js');
                 ecommerceLoaded = true;
+				this._log('loadGA', 'ecommerce');
+            }
+			
+			if (trackEnhancedEcommerce && !enhancedEcommerceLoaded) {
+                $window.__gaTracker('require', 'ec', 'ec.js');
+                enhancedEcommerceLoaded = true;
+				this._log('loadGA', 'ec');
+            }
+			
+			if (trackDisplayfeatures && !displayfeaturesLoaded) {
+                $window.__gaTracker('require', 'displayfeatures', 'displayfeatures.js');
+                displayfeaturesLoaded = true;
+				
+				for (var x = 1; i < this.trackers.length; x++) {
+					$window.__gaTracker(this.trackers[x].name + '.require', 'ecommerce', 'ecommerce.js');
+				}
+				
+				this._log('loadGA', 'displayfeatures');
             }
 
             if (trackRoutes) {
@@ -108,17 +175,39 @@ angular.module('umc-angular-google-analytics', [])
             * @private
             */
           this._trackPage = function(url,title) {
-              if (angular.isUndefined($window.__gaTracker)) { return; }
+              if (angular.isUndefined($window.__gaTracker)) {
+				return; 
+			  }
+			  
+  			  if (trackDisplayfeatures && !displayfeaturesLoaded) {
+                $window.__gaTracker('require', 'displayfeatures', 'displayfeatures.js');
+                displayfeaturesLoaded = true;
+				
+				for (var i = 1; i < this.trackers.length; i++) {
+					$window.__gaTracker(this.trackers[i].name + '.require', 'ecommerce', 'ecommerce.js');
+				}
+				
+				this._log('loadGA', 'displayfeatures');
+              }
 
               if (angular.isUndefined(url)) { url = $location.path(); }
               var fullUrl = trackPrefix + url;
               if (fullUrl !== '' && fullUrl.charAt(0) !== '/') { fullUrl = '/' + fullUrl; } //page should always start with a /
               var opts = { 'page': fullUrl };
 
-              if(angular.isUndefined(title) && angular.isDefined($rootScope.pageTitle)) { title = $rootScope.pageTitle; }
-              if(angular.isDefined(title) && title !== '') { opts.title = title; }
+              if (angular.isUndefined(title) && angular.isDefined($rootScope.pageTitle)) {
+			    title = $rootScope.pageTitle; 
+              }
+              if (angular.isDefined(title) && title !== '') { 
+			    opts.title = title; 
+			  }
 
+			  // primary
               $window.__gaTracker('send','pageview', opts);
+			  // secondary trackers
+			  for (var x = 1; x < this.trackers.length; x++) {
+				$window.__gaTracker(this.trackers[x].name + '.send','pageview', opts);
+			  }
               this._log('pageview', arguments);
           };
 
@@ -158,9 +247,10 @@ angular.module('umc-angular-google-analytics', [])
             if (angular.isUndefined($window.__gaTracker)) { return; }
 
             //guard in case ecommerce hasn't been loaded, shouldn't really happen
-            if(trackEcommerce && !ecommerceLoaded) {
+            if (trackEcommerce && !ecommerceLoaded) {
                 $window.__gaTracker('require', 'ecommerce', 'ecommerce.js');
                 ecommerceLoaded = true;
+				this._log('loadGA', 'ecommerce');
             }
 
             $window.__gaTracker('ecommerce:addTransaction', {
@@ -255,6 +345,7 @@ angular.module('umc-angular-google-analytics', [])
           // the rest of the public interface
           return {
                 _logs: me._logs,
+				trackers: me.trackers,
                 trackPage: function(url, title) {
                     // add a page event
                     me._trackPage(url, title);
@@ -279,7 +370,7 @@ angular.module('umc-angular-google-analytics', [])
                     me._trackSocial(network, action, target);
                 },
                 ga: function() {
-                    if(angular.isDefined($window.__gaTracker)) {
+                    if (angular.isDefined($window.__gaTracker)) {
                         $window.__gaTracker(arguments);
                     }
                 }