+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+<a name="0.1.8"></a>
+### 0.1.8 <2015-04-16>
+
+Allow plugin loading for each tracker.
+
+Fix track pages call for multiple trackers.
+
+Start work on enhanced ecommerce plugin support.
+
+<a name="0.1.7"></a>
+### 0.1.7 (2015-04-14)
+
+Added support for multiple tracking codes.
+
+<a name="0.1.6"></a>
+### 0.1.6 (2015-04-14)
+
+Added support for loading display features.
+
 <a name="0.1.5"></a>
 ### 0.1.5 (2014-11-10)